fix overflows in to_lookup_operands

+++ jolt-core/src/zkvm/instruction/virtual_pow2.rs
@@ -39,7 +39,7 @@ impl<const WORD_SIZE: usize> LookupQuery<WORD_SIZE> for RISCVCycle<VirtualPow2> 
 
     fn to_lookup_operands(&self) -> (u64, u128) {
         let (x, y) = LookupQuery::<WORD_SIZE>::to_instruction_inputs(self);
-        (0, x as u128 + y as u128)
+        (0, x as u128 + y as u64 as u128)
     }
 
     fn to_lookup_index(&self) -> u128 {
